wip: adding update into generators
