+++ src/generators/new/domain/entities.js
@@ -23,18 +23,47 @@ function generateEntities (from, to, level = './') {
   return requires
 }
 
-module.exports = async ({ template: { generate }, parameters: { options } }) => async () => {
+function updateEntities (entitiesPath, level = './') {
   let requires = {}
-  if (options.entities && options.entities !== true) {
-    requires = generateEntities(`../${options.entities}`, `${filesystem.cwd()}/src/domain/entities`)
-  } else {
-    await generate({
-      template: 'domain/entities/user.ejs',
-      target: 'src/domain/entities/user.js'
-    })
-    requires.User = 'require(\'./user.js\')'
-  }
+  fs.readdirSync(entitiesPath).forEach(element => {
+    const elementPath = path.join(entitiesPath, element)
+    const splittedFileName = element.split('.')
 
+    const isFile = fs.lstatSync(elementPath).isFile()
+    if (!isFile) {
+      const result = updateEntities(elementPath, `${level}${element}/`)
+      requires = Object.assign(requires, result)
+      return
+    }
+
+    // Can I require it?
+    if (isFile &&
+      splittedFileName.pop() === 'js' &&
+      splittedFileName[0] !== 'index') {
+      const entity = require(`${entitiesPath}/${element}`)
+      requires[entity.name] = `require('${level}${element}')`
+    }
+  })
+  return requires
+}
+
+module.exports = async ({ template: { generate }, parameters: { options } }, isUpdate) => async () => {
+  let requires = {}
+
+  if(isUpdate){
+    requires = updateEntities('./src/domain/entities')
+  }else{
+    if (options.entities && options.entities !== true) {
+      requires = generateEntities(`../${options.entities}`, `${filesystem.cwd()}/src/domain/entities`)
+    } else {
+      await generate({
+        template: 'domain/entities/user.ejs',
+        target: 'src/domain/entities/user.js'
+      })
+      requires.User = 'require(\'./user.js\')'
+    }  
+  }
+  
   await generate({
     template: 'domain/entities/index.ejs',
     target: 'src/domain/entities/index.js',
