+++ src/generators/new/infra/rest.js
@@ -4,7 +4,6 @@ const { objToString } = require('../../utils')
 const requireRoute = (type, entityName, useId = false) => `{ usecase: require('../../../domain/usecases/${camelCase(entityName)}/${type}${entityName}')(repositories)${useId ? ', id: \'id\'' : ''}}`
 
 module.exports = async ({ template: { generate }, filesystem }) => async () => {
-  // TODO: add condition if route
   const usecases = require(`${filesystem.cwd()}/src/domain/usecases`)
 
   // groupBy tags
